+++ src/main/java/com/nyfaria/manaunification/mixin/ManaBarOverlayMixin.java
@@ -1,0 +1,21 @@
+package com.nyfaria.manaunification.mixin;
+
+import com.llamalad7.mixinextras.injector.wrapmethod.WrapMethod;
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import io.redspace.ironsspellbooks.gui.overlays.ManaBarOverlay;
+import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.world.entity.player.Player;
+import net.minecraftforge.client.gui.overlay.ForgeGui;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(ManaBarOverlay.class)
+public class ManaBarOverlayMixin {
+
+    @Inject(method = "render", at = @At("HEAD"), cancellable = true, remap = false)
+    public void render(ForgeGui gui, GuiGraphics guiHelper, float partialTick, int screenWidth, int screenHeight, CallbackInfo ci) {
+        ci.cancel();
+    }
+}