+++ src/main/java/com/nyfaria/manaunification/mixin/MixinExample.java
@@ -1,4 +1,0 @@
-package com.nyfaria.manaunification.mixin;
-
-public class MixinExample {
-}