+++ changelog.md
@@ -1,2 +1,13 @@
-## Version 1.0.0
+# Mana Unification Changelog
+<hr>
+
+### Version 1.0.2
+- Added Apotheosis Affix and Gem Compat.
+
+### Version 1.0.1
+- Fixed Mana Not being used.
+- Fixed Channeled Spells not working.
+- Rebalance Iron Spells mana cost to be more in line with ars mana.
+
+### Version 1.0.0
 - Initial Release